+++ src/plot_page/app.py
@@ -1,6 +1,5 @@
-from dash import Dash
 import dash_bootstrap_components as dbc
-
+from dash import Dash
 
 app = Dash(__name__, external_stylesheets=[dbc.themes.BOOTSTRAP])
 app.title = "Analyse Dash App"
