+++ src/plot_page/view/pages/plot_2d.py
@@ -1,20 +1,12 @@
-import base64
-import datetime
-import io
-import json
-import os
-from typing import Any, List, Tuple, Union
-from urllib.parse import quote as urlquote
 
 import dash
 import dash_bootstrap_components as dbc
-import pandas as pd
-import plotly.express as px
-import plotly.graph_objects as go
-from dash import Dash, Input, Output, State, callback, dash_table, dcc, html
+from dash import Input, Output, State, dash_table, dcc, html
 from dash.dependencies import Input, Output
+
 from plot_page.app import app
-from plot_page.control.data_operations import filter_columns
+from plot_page.control.data_operations import check_line_config, filter_columns
+from plot_page.control.plot_functions import plot_data
 from plot_page.view.components.default_component import get_upload_component
 
 
@@ -48,103 +40,6 @@ def graph_setting() -> dbc.Card:
             ),
         ]
     )
-
-
-#####################################################################################################################################################
-def plot_line(fig: go.Figure, splitted_data: dict[str, pd.DataFrame], x_axis: str, y_axis: str, settings: dict) -> None:
-    """Add the line to the figure.
-
-    Args:
-        fig (go.Figure): The existing figure.
-        splitted_data (dict[str, pd.DataFrame]): Dictionary of key and dataframe pair.
-        x_axis (str): The selected attribute for the x_axis.
-        y_axis (str): The selected attribute for the y_axis.
-        settings (dict): The configuration of the line.
-    """
-    if settings["plot_kind"] == "History":
-        for key, val in splitted_data.items():
-            fig.add_trace(go.Scatter(x=val[x_axis], y=val[y_axis], mode="lines", name=f"trace_{key}"))
-
-    if settings["plot_kind"] == "Min":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].min()
-            fig.add_trace(go.Scatter(x=plot_data.index.values, y=plot_data, mode="lines", name=f"min_{key}"))
-
-    if settings["plot_kind"] == "Max":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].max()
-            fig.add_trace(go.Scatter(x=plot_data.index.values, y=plot_data, mode="lines", name=f"min_{key}"))
-
-    if settings["plot_kind"] == "Median":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].median()
-            fig.add_trace(go.Scatter(x=plot_data.index.values, y=plot_data, mode="lines", name=f"median_{key}"))
-
-    if settings["plot_kind"] == "Mean":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].mean()
-            fig.add_trace(go.Scatter(x=plot_data.index.values, y=plot_data, mode="lines", name=f"mean{key}"))
-
-
-#####################################################################################################################################################
-def plot_bar(fig: go.Figure, splitted_data: dict[str, pd.DataFrame], x_axis: str, y_axis: str, settings: dict) -> None:
-    """Add the line to the figure.
-
-    Args:
-        fig (go.Figure): The existing figure.
-        splitted_data (dict[str, pd.DataFrame]): Dictionary of key and dataframe pair.
-        x_axis (str): The selected attribute for the x_axis.
-        y_axis (str): The selected attribute for the y_axis.
-        settings (dict): The configuration of the line.
-    """
-    if settings["plot_kind"] == "History":
-        for key, val in splitted_data.items():
-            fig.add_trace(go.Bar(x=val[x_axis], y=val[y_axis], name=f"trace_{key}"))
-
-    if settings["plot_kind"] == "Min":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].min()
-            fig.add_trace(go.Bar(x=plot_data.index.values, y=plot_data, name=f"min_{key}"))
-
-    if settings["plot_kind"] == "Max":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].max()
-            fig.add_trace(go.Bar(x=plot_data.index.values, y=plot_data, name=f"min_{key}"))
-
-    if settings["plot_kind"] == "Median":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].median()
-            fig.add_trace(go.Bar(x=plot_data.index.values, y=plot_data, name=f"median_{key}"))
-
-    if settings["plot_kind"] == "Mean":
-        for key, val in splitted_data.items():
-            plot_data = val.groupby(by=x_axis)[y_axis].mean()
-            fig.add_trace(go.Bar(x=plot_data.index.values, y=plot_data, name=f"mean{key}"))
-
-
-#####################################################################################################################################################
-def plot_data(data_table: dict[str, list[dict]], plot_setting: list[dict], x_axis: str, y_axis: str) -> dcc.Graph:
-    """Plot the selected tables ans settings.
-
-    Args:
-        data_table (dict[str, list[dict]]): All table data.
-        plot_setting (list[dict]): The selected plot settings.
-        x_axis (str): The selected x_axis attribute.
-        y_axis (str): The selected y_axis attribute.
-
-    Returns:
-        dcc.Graph: The dcc.Graph that should be plotted.
-    """
-    fig = go.Figure()
-    for val in plot_setting:
-        if val["type"] == "Line":
-            splitted_data = split_data(data_table, val["group_attributes"])
-            plot_line(fig, splitted_data, x_axis, y_axis, val)
-        if val["type"] == "Bar":
-            splitted_data = split_data(data_table, val["group_attributes"])
-            plot_bar(fig, splitted_data, x_axis, y_axis, val)
-
-    return dcc.Graph(figure=fig)
 
 
 #####################################################################################################################################################
@@ -200,26 +95,6 @@ def plot_config() -> dbc.Card:
             html.Div(dash_table.DataTable(data=[], id="plot_settings", page_size=20), style={"padding": "10px"}),
         ]
     )
-
-
-#####################################################################################################################################################
-def check_line_config(plot_type: str, value_to_plot: str, group_by: list[str]) -> dict:
-    """Check if the line configuration is ok.
-
-    Args:
-        plot_type (str): The type that should be plotted.
-        value_to_plot (str): Extended information about what should be plotted.
-        group_by (list[str]): List of attributes the data should be grouped by.
-
-    Returns:
-        dict: Dictionary that contains all informations
-    """
-    if plot_type is None:
-        return None
-    if value_to_plot is None:
-        return None
-
-    return {"type": plot_type, "plot_kind": value_to_plot, "group_attributes": group_by}
 
 
 #####################################################################################################################################################
@@ -400,6 +275,7 @@ def update_selectable_data(table_data: dict[str, list[dict]]) -> list[str]:
         return []
     return list(table_data)
 
+
 #####################################################################################################################################################
 @app.callback(Output("plot_config_table", "data"), Input("select_2d_table", "value"))
 def reset_config_table(data: dict) -> dict:
@@ -412,7 +288,8 @@ def reset_config_table(data: dict) -> dict:
         dict: The new plot_config table.
     """
     return {"settings": []}
-              
+
+
 #####################################################################################################################################################
 @app.callback(Output("selected_table_data", "data"), Input("select_2d_table", "value"), State("table_data", "data"))
 def selected_table(selected_tables: list[str], table_data: dict[str, list[dict]]) -> dict[str, list[dict]]:
@@ -430,40 +307,3 @@ def selected_table(selected_tables: list[str], table_data: dict[str, list[dict]]
     if table_data is None:
         return dash.no_update
     return {key: table_data[key] for key in selected_tables}
-
-
-#####################################################################################################################################################
-def grouping_pd(group: str, data: dict[str, pd.DataFrame]) -> dict[str, pd.DataFrame]:
-    """Group the dataframe by the attribute values.
-
-    Args:
-        group (str): Attribute name the data should be grouped with.
-        data (dict[str, pd.DataFrame]): The already grouped dataframes.
-
-    Returns:
-        dict[str, pd.DataFrame]: The new grouped dataframe.
-    """
-    groups = list(set(data[list(data.keys())[0]][group].tolist()))
-    return {f"{key}_{attribute}": val[getattr(val, group) == attribute] for key, val in data.items() for attribute in groups}
-
-
-#####################################################################################################################################################
-def split_data(data: dict, grouping: list[str]) -> dict:
-    """Split dictionary of data in groups.
-
-    Args:
-        data (dict): The dictionary that contaisn the data.
-        grouping (list[str]): List of attributes the data should be grouped.
-
-    Returns:
-        dict: The resulting splitted dataset.
-    """
-    res = {key: pd.DataFrame.from_dict(val) for key, val in data.items()}
-
-    if grouping is None:
-        return res
-
-    for g in grouping:
-        res = grouping_pd(g, res)
-
-    return res
